feat(refactor): registration refactor began, may continue to incorporate loops
